Fix allocation of interfaces to netns on reboot
In newer versions of ip there are also lines that report altname for
interfaces as well. The line in the systemd unit was not handling this
well as it was expecting always two lines per interface.
ip now comes with an option to have brief output which can be parsed in
a better way.

Signed-off-by: Daniel Failing <daniel.failing@cern.ch>
Related-Bug: #2027995
Change-Id: I90cd69f9332dc1381438e78ea8af8e79b95c412d