Fix functional tests due to oslo.context change
Oslo.context recently[1] changed the RequestContext to not explictly have a
project_id in the base object.
This broke the Octavia function tests as it was mocking this object for
api test cases.
This patch changes how the Octavia functional tests are mocking the
project ID in the api functional tests.

[1] https://review.opendev.org/c/openstack/oslo.context/+/894622

Change-Id: I4c536632ead468080d40525e77978d536f94b3cc