devstack: Omit flavor create --id auto parameter
This alias has been broken in OSC since it switched flavor commands to
SDK in OSC 5.5.0 (March 2021), resulting in flavors being created with
an ID of 'auto' instead. While this is now being fixed, the option has
never been necessary as Nova will automatically configure this for you,
so we can simply remove it instead.

Change-Id: Ic379eb7f41fd722e0574f2427b328b4095fc08ed
Signed-off-by: Stephen Finucane <stephenfin@redhat.com>
Related-bug: #2120833