Fix Amphora controller IP list update
Updating the Amphora agent configuration was not triggering the health
manager sender to reload it's configuration file. This meant that it
would not adopt the updated controller IP and port list.
This patch fixes this by signalling the health manager sender process
via a multiprocessing manager queue that it needs to reload it's
configuration.

Closes-Bug: #2121691

Change-Id: I1c8520745bf1bb59d483cc7761929794c58e5021
Signed-off-by: Michael Johnson <johnsomor@gmail.com>