API data mapping performance enhancements
This patch aims to dramatically improve API performance by bypassing
the data_model layer (only at the API) and instead mapping directly
from the DB models to the API model by providing an ORM-based
alternative processing path.

On the same DevStack with the same data, a loadbalancer show on a large
loadbalancer was observed to be up to 99% faster with this patch over
master.

Includes a re-alignment of Quotas API controller and types.

Also includes a "revert" of the recursion_depth patch as now obsolete.

Closes-Bug: #2134543

Change-Id: I23cbde2ff90c4a261bcb9007f36954347d17c719
Signed-off-by: Théo Van Gyzel <theo.van-gyzel@ovhcloud.com>