Fix spurious logging message in amphora-agent API
The details endpoint for the amphora-agent API checks for a tuned profile. If the file doesn't exist it logs a message. Since tuned isn't required it's very likely this file doesn't exist and this leads to a lot of misleading log messages. I've updated the endpoint to suppress the logging in the event the file doesn't exist, any other exceptions will still generate the log message. Unit tests were also updated using Claude code to add a test case where the file doesn't exist

Change-Id: I7f1d423e6464a70c9546d92052ec86cdcdd487f9
Signed-off-by: Richard Cruise <rcruise@redhat.com>
Assissted-by: Claude code <https://www.anthropic.com>